move native_tls stuff to separate file module
